Add components folder and moved app.js in components
